Fix .tbss placement check bug
Fix the bug that ld would complain if the VMA of .tbss ends up outside
the PT_LOAD segment.  That's totally irrelevant, since the VMA of the
.tbss section is never used anyway.  There was already a fix to ignore
the length of the .tbss section in the context of PT_LOAD segments,
so extend that to completely ignore VMA in that case.